📈 perf(img): viewport img size calc func moved to util & StaticImg gridnum sizes added
StaticImg applied to blog post thumbnail & GridNum default values adjusted to not render full in
tablet and above by default
