+++ src/components/Image/StaticImage.tsx
@@ -1,6 +1,5 @@
 import Image from 'next/image';
-import * as React from 'react';
-import { calcStdImageWidth, cn } from '@/src/util/utils';
+import { cn, calcStdImageWidth, dynamicViewport } from '@/src/util/utils';
 
 interface StaticImageProps {
   src: string;
@@ -10,6 +9,7 @@ interface StaticImageProps {
   targetHeight?: number;
   quality?: number;
   aspectRatio?: string;
+  gridNums?: number[];
   isPriority?: boolean;
   wrapperClassName?: string;
   imageClassName?: string;
@@ -23,6 +23,7 @@ const StaticImage = ({
   targetHeight = 1080,
   quality,
   aspectRatio = 'aspect-auto',
+  gridNums = [1, 2, 2],
   isPriority = false,
   wrapperClassName,
   imageClassName,
@@ -36,6 +37,7 @@ const StaticImage = ({
       quality={quality ?? 75}
       width={calcStdImageWidth(width, height)} // 다른 dimension 입력하면 1080 height 기준으로 맞춤
       height={targetHeight}
+      sizes={dynamicViewport(gridNums)}
       priority={isPriority}
       className={cn('relative', imageClassName)}
     />
